--- app/(home)/layout.tsx
@@ -1,23 +1,0 @@
-import '../globals.css';
-import { Inter } from 'next/font/google';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata = {
-	title: 'Ian Payntar Smith',
-	description: '',
-};
-
-export default function RootLayout({
-	children,
-}: {
-	children: React.ReactNode;
-}) {
-	return (
-		<html lang="en">
-			<body className={inter.className}>
-				<main>{children}</main>
-			</body>
-		</html>
-	);
-}
